requirements.txt
Django==5.2.0
djangorestframework==3.15.1
django-filter==24.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
whitenoise==6.6.0